Created SILC Crypto Toolkit git repository.
The new git repository contains the entire history of the SILC
crypto library, math library, etc., now named the SILC Crypto Toolkit.
In addition of that it contains the entire history of the SILC
development since the crypto stuff has been part of that from the
very beginning.  None of the history has been removed.  Only things
that have been removed are unneeded branches and tags.

In this changeset we remove all unnecessary files and directories and
make the tree compile.

The SILC Crypto Toolkit now uses the SILC Runtime Toolkit and cannot
be compiled without first installing SRT.